finish %u

+++ libft/ft_nbrlen.c
@@ -6,11 +6,11 @@
 /*   By: mirsella <mirsella@protonmail.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 18:31:57 by mirsella          #+#    #+#             */
-/*   Updated: 2022/11/22 18:32:16 by mirsella         ###   ########.fr       */
+/*   Updated: 2022/11/22 22:56:41 by mirsella         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_nbrlen(int n)
+int	ft_nbrlen(long long n)
 {
 	int	len;
 
@@ -30,7 +30,7 @@ int	ft_nbrlen(int n)
 	return (len);
 }
 
-int	ft_nbrlen_base(int n, int base)
+int	ft_nbrlen_base(long long n, int base)
 {
 	int	len;
